refactor(coral): move recipe and examples under beta
