+++ recipes/__init__.py
@@ -4,6 +4,7 @@
 weight methods) its ``slime/`` backend half, plus ``examples/`` — the
 runnable Harbor task + harness stacks that drive it. ``recipes/basic`` is the
 example for the core record-only ``recipe``, which has no method package.
+Beta methods live under ``recipes/beta/<method>/`` with the same layout.
 Reef does not import this tree, and none of it ships in the wheel. Example
 deployments select recipe and loss classes through explicit dotted references.
 """
