+++ recipes/beta/coral/examples/coral_demo/pyproject.toml
@@ -19,4 +19,4 @@ requires = ["setuptools>=68"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
-include = ["recipes.coral*"]
+include = ["recipes.beta.coral*"]
