+++ recipes/beta/coral/examples/coral_demo/smoke/scripted_agent.py
