+++ recipes/beta/coral/examples/coral_demo/task/litellm_config.yaml
