+++ recipes/beta/coral/examples/coral_demo/task/seed/solution.py
