+++ recipes/beta/coral/examples/coral_demo/task/task.yaml
