+++ recipes/beta/coral/recipe.py
@@ -1,6 +1,6 @@
 """The CORAL weight-training recipe: what a deployment names in ``serve.yaml``.
 
-Binds :class:`~reef_coral.processor.CoralProcessor` to the grouped
+Binds :class:`~recipes.beta.coral.processor.CoralProcessor` to the grouped
 relative-reward training machinery. CORAL sibling groups have the same shape
 as TTT-Discover steps — comparison sets of scored rollouts from one shared
 starting point — so the recipe reuses the ``tttd`` step preparer (grouped
@@ -8,9 +8,9 @@ adaptive-entropic leave-one-out advantages) and the ``tttd`` Slime loss
 family rather than duplicating either. The recipe's own configuration is the
 group barrier.
 
-Deployment reference: ``recipes.coral.recipe:CoralRecipe`` (after
-``pip install -e recipes/coral`` alongside the repository cookbook, which
-provides the ``tttd`` registrations this recipe reuses).
+Deployment reference: ``recipes.beta.coral.recipe:CoralRecipe``. Run from
+the repository checkout, which also provides the ``tttd`` registrations this
+recipe reuses; ``reef serve`` finds it beside the example config.
 """
 
 from __future__ import annotations
@@ -20,7 +20,7 @@ from dataclasses import dataclass
 # The tttd step preparer and loss-family reference live in the repository
 # cookbook; importing them is what registers both names this recipe binds.
 import recipes.tttd  # noqa: F401  (registration side effect)
-from recipes.coral.processor import CoralProcessor
+from recipes.beta.coral.processor import CoralProcessor
 from reef.recipe.base import WeightTrainingRecipe, WeightTrainingSpec
 from reef.recipe.config_fields import config_field
