+++ recipes/beta/coral/reporter.py
@@ -14,7 +14,7 @@ from collections.abc import Mapping
 from dataclasses import dataclass, field
 from typing import Any
 
-from recipes.coral.journal import deterministic_report_id
+from recipes.beta.coral.journal import deterministic_report_id
 
 logger = logging.getLogger(__name__)
